Renamed scraper module to spatula.py and script to run_spatula.py, updated readme.md
